Add and use TranslatableTrait in requests
