+++ src/Request/GetSceneRequest.php
@@ -3,6 +3,7 @@
 namespace Flowly\Content\Request;
 
 use Flowly\Content\Request\Part\ResolutionTrait;
+use Flowly\Content\Request\Part\TranslatableTrait;
 use Symfony\Component\Validator\Constraints as Assert;
 
 /**
@@ -11,17 +12,13 @@ use Symfony\Component\Validator\Constraints as Assert;
 class GetSceneRequest
 {
     use ResolutionTrait;
+    use TranslatableTrait;
 
     /**
      * @var string
      * @Assert\Uuid(message="Parameter 'id' must be valid uuid.")
      */
     private string $id;
-
-	/**
-	 * @var string|null
-	 */
-	private ?string $language = null;
 
     public function __construct(string $id) { $this->id = $id; }
 
@@ -40,22 +37,6 @@ class GetSceneRequest
     {
         $this->id = $id;
     }
-
-	/**
-	 * @return string|null
-	 */
-	public function getLanguage(): ?string
-	{
-		return $this->language;
-	}
-
-	/**
-	 * @param string|null $language
-	 */
-	public function setLanguage(?string $language): void
-	{
-		$this->language = $language;
-	}
 
     public function toArray(): array
     {
