+++ src/Request/GetSceneSuggestRequest.php
@@ -5,6 +5,7 @@ namespace Flowly\Content\Request;
 use Flowly\Content\Request\Part\LinksTrait;
 use Flowly\Content\Request\Part\OrderTrait;
 use Flowly\Content\Request\Part\ResolutionTrait;
+use Flowly\Content\Request\Part\TranslatableTrait;
 use Symfony\Component\Validator\Constraints as Assert;
 
 class GetSceneSuggestRequest
@@ -12,6 +13,7 @@ class GetSceneSuggestRequest
     use OrderTrait;
     use LinksTrait;
     use ResolutionTrait;
+    use TranslatableTrait;
 
     /**
      * @var string
